fix: align dex price on mobile
To calculate using value of b with ratio ba
